requirements.txt
# Defense Intelligent Agent Platform - 필수 패키지
# 최종 업데이트: 2026-01-17
# Python 3.9 이상 (3.12 이상 권장)

# ============================================
# 핵심 웹 프레임워크 (FastAPI & API)
# ============================================
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.6
pydantic>=2.0.0

# ============================================
# 데이터 분석 및 수치 연산
# ============================================
pandas>=2.0.0
openpyxl>=3.1.0  # Excel 파일 읽기/쓰기
numpy>=1.24.0
scipy>=1.10.0    # 고급 수치 연산 및 기하학 계산
pyarrow>=10.0.0  # Pandas 성능 향상

# ============================================
# 시각화 및 GIS
# ============================================
plotly>=5.18.0
networkx>=3.0    # 그래프 분석
matplotlib>=3.7.0  # 기본 시각화
graphviz>=0.20.1  # 그래프 시각화 (시스템 설치 필요)
pyvis>=0.3.2      # 대화형 네트워크 그래프

# ============================================
# 온톨로지 및 세맨틱 기술
# ============================================
rdflib>=7.0.0     # RDF 그래프 처리
owlrl>=6.0.2      # OWL-RL 추론 엔진
owlrl>=6.0.0      # OWL 추론 엔진

# ============================================
# RAG (Retrieval Augmented Generation) & 벡터 검색
# ============================================
faiss-cpu>=1.12.0
# faiss-gpu>=1.12.0  # GPU 사용 시 전환 필요
sentence-transformers>=2.2.0
scikit-learn>=1.0.0

# ============================================
# LLM 및 생성형 AI (선택적)
# ============================================
openai>=1.0.0
transformers>=4.35.0
torch>=2.7.0
# GPU 버전 설치 시: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
accelerate>=0.24.0

# ============================================
# 문서 처리 및 보고서 생성 (docs/reports)
# ============================================
pdfplumber>=0.10.0
python-docx>=1.0.0
jinja2>=3.1.0

# ============================================
# 유틸리티 및 설정
# ============================================
pyyaml>=6.0
requests>=2.28.0
python-dateutil>=2.8.2
tqdm>=4.65.0

# ============================================
# 설치 및 실행 주의사항
# ============================================
# 1. Graphviz: 시스템에 Graphviz 소프트웨어가 설치되어 있어야 함
# 2. 경로 문제: Windows 환경에서 venv 경로 오류 발생 시, 
#    'uvicorn' 대신 'python -m uvicorn' 명령어를 사용 권장
# 3. 브라우저: 최신 Chrome 또는 Edge 브라우저 권장

